IntelCusEngProj/requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysqlclient==2.2.0
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
protobuf==4.21.12
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
